setup the steps of refactor
