+++ Makefile
@@ -2,10 +2,13 @@ download:
 	@echo "downloading data..."
 	@echo curl https://data.cityofnewyork.us/api/views/4p5c-cbgn/rows.csv?accessType=DOWNLOAD --output data/data.csv
 
-install | i:
+install:
 	@echo "installing dependencies..."
 	@poetry install --only main
 
+idev:
+	@echo "installing development dependencies..."
+	@poetry install --only dev
 lint:
 	@echo "lint checking..."
 	@flake8
@@ -13,3 +16,7 @@ lint:
 test:
 	@echo "running test cases..."
 	@nose2 --config=nose2.cfg -v -F
+
+coverage:
+	@echo "running test coverage..."
+	@coverage report ./test/*_test.py
