+++ functions/__init__.py
@@ -1,0 +1,9 @@
+from functions.first import compute_and_add_trip_duration
+from functions.second import add_hour_of_day_and_day_of_week
+from functions.third import compute_predictions
+from functions.fourth import get_predictions
+
+__all__ = (['compute_and_add_trip_duration',
+            'add_hour_of_day_and_day_of_week',
+            'compute_predictions',
+            'get_predictions'])
