+++ test/first_test.py
@@ -1,0 +1,6 @@
+from unittest import TestCase
+
+
+class Test(TestCase):
+    def test_compute_and_add_trip_duration(self):
+        pass
