+++ test/fourth_test.py
@@ -1,0 +1,6 @@
+from unittest import TestCase
+
+
+class Test(TestCase):
+    def test_get_predictions(self):
+        pass
