+++ test/second_test.py
@@ -1,0 +1,6 @@
+from unittest import TestCase
+
+
+class Test(TestCase):
+    def test_add_hour_of_day_and_day_of_week(self):
+        pass
